Flesh out `pub` scenario
1. Add a vulnerable package version
2. Add a `dev_dependencies` section
3. Adding these two packages also pulls in some transitive packages,
   which we were previously lacking.

I regenerated the lockfile using the following steps:
1. `docker pull dart`
2. `docker run -it -v /Users/jeffwidman/Code/github/smoke-tests:/root dart`
3. Within the container: `dart pub get`

This resulted in generating a new updated lockfile.

Note that because so many additional transitive deps were pulled in,
this will take a little longer. However I tested this on our staging
repo and the entire pub run took ~82 seconds, so we should be okay.