Merge the cache-one and cache-all workflows
This merges the cache-one and cache-all workflows.

There was no functional difference between them, so by adding a single
step we can make it so it supports both caching one and caching all...

Makes it so that it doesn't go out of sync.

I will be following a similar strategy for a new workflow that I'm
working on that regenerates the smoke tests and then opens a PR so that
we don't have to do it locally but instead can just run an action to
regenerate the test output, so this is an initial stab at getting this
type of workflow going.