Regenerate Bundler smoke test results
New version of Bundler 2.5 makes slightly different network requests
that cause our smoke tests cached requests to not be hit, which causes
some versions to be bumped further than our expectations.

This is not a problem in Bundler or Dependabot, so easiest way to
unblock the upgrade to Bundler 2.5 is to update smoke test expectations.