Upgrade Bundler smoke test for Ruby 3.3.6
Regenerate three Bundler smoke tests

These tests start failing with Ruby 3.3.6 because it provides a higher
version of json by default than the recorded API responses, so that's
what Dependabot resolves to.

I believe the easiest way to fix that is to regenerate the tests with
updated cache, so that the above situation no longer happens.